Unit conversions
A new field called "conversions:" has been added to the yaml config file. Here you can add, for each get_all parameter, the conversion factor to apply.
This must also be done to the parameters set to the ESP (insp. press and supp. pressure for now). For those, there is an additional field called "conversion:", which has to convert in the opposite way in this case (cmH2O to mbar)