Initially set Settings to what is in the ESP (if ESP is running)
If the ESP is running, read the settings params from the ESP and update the values in the Settings panel.
If the ESP is NOT running, don't do this, but follow the classic behaviour. This is because if we click on "Continue Ventilation" we don't want to read the default parameters in the ESP, but we want to acttively set the previously saved params to the ESP. And if we click on "New patient" we want to start from fresh parameters as well.