adding a pylintrc file
the file has been generated with PyLint 2.4.4 by:
        pylint --generate-rcfile > .pylintrc
and then adding PyQt5 in the extension package whitelist. This because
by default PyLint doesn't load C extensions for safety reasons
(arbitrary code execution).

The file has been put in this folder for current convenience. Can be
moved to the top directory if needed.